+++ packages/webapp/src/stories/Pages/Intro/Signup3/Signup3.stories.js
@@ -1,8 +1,10 @@
 import React from 'react';
 import Signup3 from './Signup3';
+import decorators from '../components/decorators';
 
 export default {
   title: 'Form/Intro/Signup3',
+  decorators: decorators,
   component: Signup3,
 };
