+++ packages/webapp/src/stories/Pages/Intro/Signup6/Signup6.stories.js
@@ -1,8 +1,10 @@
 import React from 'react';
 import Signup6 from './Signup6';
+import decorators from '../components/decorators';
 
 export default {
   title: 'Form/Intro/Signup6',
+  decorators: decorators,
   component: Signup6,
 };
