+++ packages/webapp/src/stories/Pages/Intro/Signup7/Signup7.stories.js
@@ -1,8 +1,10 @@
 import React from 'react';
 import Signup7 from './Signup7';
+import decorators from '../components/decorators';
 
 export default {
   title: 'Form/Intro/Signup7',
+  decorators: decorators,
   component: Signup7,
 };
